bundes-db-init/requirements.txt
pymongo
Pillow
pandas
mysql-connector-python
